# project2-recipes/server.js
const express = require('express');
const session = require('express-session');
const { initDb } = require('./db/connect');
require('dotenv').config();

const app = express();
const port = process.env.PORT || 3000;

app.use(express.json());

// express.json() throws if the body is not valid JSON
app.use((err, req, res, next) => {
  if (err instanceof SyntaxError && err.status === 400) {
    return res.status(400).json({
      message: 'Validation failed.',
      errors: ['Request body is not valid JSON.']
    });
  }
  next(err);
});

// Render sits behind a proxy. Without this the secure cookie never gets set.
app.set('trust proxy', 1);

app.use(
  session({
    secret: process.env.SESSION_SECRET || 'local-dev-secret',
    resave: false,
    saveUninitialized: false,
    cookie: {
      // Only send the cookie over https once we are deployed
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'lax',
      maxAge: 1000 * 60 * 60 * 24 // one day
    }
  })
);

// Passport has to come after the session, because it stores the user there
const passport = require('./config/passport');
app.use(passport.initialize());
app.use(passport.session());

app.use((req, res, next) => {
  res.setHeader('Access-Control-Allow-Origin', '*');
  res.setHeader('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Z-Key');
  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
  next();
});

app.use('/', require('./routes'));

// Nothing above matched
app.use((req, res) => {
  res.status(404).json({ message: 'Route not found: ' + req.method + ' ' + req.originalUrl });
});

// Catches anything the controllers missed
app.use((err, req, res, next) => {
  console.error('Unhandled error:', err);
  res.status(500).json({ message: 'Something went wrong on the server.', error: err.message });
});

// Connect to the database first, then start listening
initDb()
  .then(() => {
    app.listen(port, () => {
      console.log('Web server is listening on port ' + port);
    });
  })
  .catch((err) => {
    console.error('Failed to connect to the database:', err);
    process.exit(1);
  });