# project2-recipes/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recipe Box API",
    "description": "CSE 341 Project 2. A REST API for recipes and the chefs who wrote them. Two MongoDB collections with full CRUD, field validation on every write, and error handling on every route.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "paths": {
...
    "/recipes/": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get all recipes",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Recipes"
        ],
        "summary": "Create a new recipe",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not logged in"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "githubOAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recipe"
              },
              "example": {
                "title": "Blueberry Pancakes",
                "description": "Fluffy buttermilk pancakes folded with fresh blueberries.",
                "cuisine": "American",
                "difficulty": "easy",
                "prepMinutes": 10,
                "cookMinutes": 15,
                "servings": 4,
                "ingredients": [
                  "200g flour",
                  "2 eggs",
                  "300ml buttermilk",
                  "150g blueberries"
                ],
                "instructions": [
                  "Whisk the dry ingredients.",
                  "Fold in the wet ingredients.",
                  "Cook on a medium griddle.",
                  "Flip once."
                ],
                "chefId": "000000000000000000000c01"
              }
            }
          }
        }
      }
    },
    "/recipes/{id}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get a single recipe by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The 24 character id of the document",
            "schema": {
              "type": "string",
              "default": "000000000000000000000a01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Recipes"
        ],
        "summary": "Update a recipe by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The 24 character id of the document",
            "schema": {
              "type": "string",
              "default": "000000000000000000000a01"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not logged in"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "githubOAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recipe"
              },
              "example": {
                "title": "Spaghetti Carbonara",
                "description": "A Roman pasta dish, updated to serve six people.",
                "cuisine": "Italian",
                "difficulty": "hard",
                "prepMinutes": 12,
                "cookMinutes": 18,
                "servings": 6,
                "ingredients": [
                  "600g spaghetti",
                  "220g guanciale",
                  "6 egg yolks",
                  "120g pecorino romano",
                  "black pepper"
                ],
                "instructions": [
                  "Boil the pasta in salted water.",
                  "Crisp the guanciale.",
                  "Whisk yolks with pecorino.",
                  "Toss off the heat."
                ],
                "chefId": "000000000000000000000c01"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipes"
        ],
        "summary": "Delete a recipe by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The 24 character id of the document",
            "schema": {
              "type": "string",
              "default": "000000000000000000000a01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not logged in"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "githubOAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "githubOAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Log in by opening /login in a browser tab. GitHub sends you back with a session cookie, and Swagger then includes that cookie automatically. Without it every POST, PUT and DELETE returns 401."
      }
    },
    "schemas": {
      "Recipe": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "required": {
            "type": "array",
            "example": [
              "title",
              "description",
              "cuisine",
              "difficulty",
              "prepMinutes",
              "cookMinutes",
              "servings",
              "ingredients",
              "instructions",
              "chefId"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "title": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Spaghetti Carbonara"
                  },
                  "description": {
                    "type": "string",
                    "example": "2-100 characters"
                  }
                }
              },
              "description": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "A Roman pasta dish built on eggs, cheese and cured pork."
                  },
                  "description": {
                    "type": "string",
                    "example": "10-500 characters"
                  }
                }
              },
              "cuisine": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Italian"
                  },
                  "description": {
                    "type": "string",
                    "example": "2-50 characters"
                  }
                }
              },
              "difficulty": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "enum": {
                    "type": "array",
                    "example": [
                      "easy",
                      "medium",
                      "hard"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "type": "string",
                    "example": "medium"
                  }
                }
              },
              "prepMinutes": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 10
                  },
                  "description": {
                    "type": "string",
                    "example": "Whole number, 0-1440"
                  }
                }
              },
              "cookMinutes": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 15
                  },
                  "description": {
                    "type": "string",
                    "example": "Whole number, 0-1440"
                  }
                }
              },
              "servings": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 4
                  },
                  "description": {
                    "type": "string",
                    "example": "Whole number, 1-100"
                  }
                }
              },
              "ingredients": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "example": {
                    "type": "array",
                    "example": [
                      "400g spaghetti",
                      "150g guanciale"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "description": {
                    "type": "string",
                    "example": "At least one item"
                  }
                }
              },
              "instructions": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "example": {
                    "type": "array",
                    "example": [
                      "Boil the pasta.",
                      "Toss off the heat."
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "description": {
                    "type": "string",
                    "example": "At least one item"
                  }
                }
              },
              "chefId": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "6ab02364ab12e572f6ca9fd5"
                  },
                  "description": {
                    "type": "string",
                    "example": "Must match an existing chef"
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}